[Workflows][US_IA] Add new firestore types for Action History (Recidiviz/recidiviz-dashboards#8261)
Adds ActionHistory and other relevant types for the firestore OpportunityUpdate type.

Relevant design doc section is here. We need a way to store a timeline of various officer actions and supervisor responses in order to calculate where a given opportunity is currently located in the IA ED flow. We'll be following up with a task to add update functions that actually interact with firestore.

GitOrigin-RevId: 8033710695d66ac92a852b37111919ec109281e8