[Workflows][US_IA] Add logic to IA opportunity firestore functions (Recidiviz/recidiviz-dashboards#8291)
Adds logic to existing function stubs on IA opp.

Updates:

setOfficerAction to append the given action to the Action History and call the firestore update function
setSupervisorResponse to add the given supervisor response to the most recent officer action and call the update function
deleteActionHistory to call the delete function.

GitOrigin-RevId: 433a4302a8eb39308d3e29712ace002a83955feb