[Workflows][US_IA] Add conditional items to the criteria list (Recidiviz/recidiviz-dashboards#8378)
The goal of this ticket was to handle a non-disqualifying criteria around victim contact that for stakeholder/policy reasons needs to be prioritized in the UI in the same manner as typical criteria.

Implementation details:
We needed to modify both the opportunity config in the admin panel and the IA ED opp schema/class.

In the "Data Requirements to Check" section, we wanted to display "Registered victim has been contacted" when the victim flag is set on the record:

added a nonOmsCriteria field with handlebars syntax to the opportunity config in the admin panel:
{{#if record.victimFlag}}Registered victim has been contacted{{/if}}
filtered out empty string nonOMSRequirements in the IA opportunity class.
In the "Validated by ICON" section, we wanted to switch strings based on the victim flag being set.

override the requirementsMet section to display our strings conditionally...

GitOrigin-RevId: b6fc0c4ccae05afa20b0bb0fa69295ca49662136