[Tasks][US_TX] Bug bash follow-ups (Recidiviz/recidiviz-dashboards#8396)
Fix an issue where absolute positioning was causing the snooze task meatball menu to display improperly on Safari. Though we had detected this as a mobile issue on iOS, I was able to replicate the original error on Safari on my computer, so you can test by opening the preview app on Safari.

Format dates consistently (Oct 31, 2024 instead of 10/31/24)

Narrow the Tasks V2 page to 64% of the viewport width so that the search bar, table view toggle, and list view items fit alongside the sidebar (at 100% zoom on a typically sized laptop screen).

GitOrigin-RevId: 513c38e786a0d2cb5d13a7d93bb0abc5289a546b