[Workflows][US_UT] Determine "Report Submitted" based on opportunity metadata (Recidiviz/recidiviz-dashboards#8602)
For Utah's Early Termination opportunity, TTs were confused that the Report Submitted tab did not reflect report submissions in their OMS.
This recidiviz-data PR introduces a new tabName value, REPORT_SUBMITTED, for people who have had a report submitted tin the past few months.

This PR updates the frontend to disable marking people as "submitted" through the UI and instead only put people in the Report Submitted tab if they have that designation, and adds a new client/opportunity fixture to test this case. This PR also fixes a minor bug: opportunities that have a submitted update in clientUpdates should no longer be treated as submitted if their config says they don't support that status.

It is possibly still confusing that users can mark the Report Submitted people ineligible, but since conditionally enabling the Mark Ineligible button is not default behavior right now, I think easier to go with this and update it if that's confusing than to prematurely cover that case.

GitOrigin-RevId: 77595f9307875f5461f9a643201ed0d3dd225776