[Workflows][US_IA] Filter out empty criteria copy on OppBase (Recidiviz/recidiviz-dashboards#8608)
Filter out requirements with empty text from the requirementsMet getter

We made a similar change previously for conditional non-OMS requirements. We're back again, because IA now has a conditional eligibility criteria - the corresponding handlebars change has already been made in the opp config, and we now need to filter out that empty string so we don't see an empty criteria item in the UI. This is a very similar change to one made previously for the nonOmsRequirements getter (reference).

Criteria in question: "Is not serving a lifetime sentence"

GitOrigin-RevId: 61528b8183ac030892a0044bab356fc738265585