[US_AZ][Workflows] Always display ACIS TPR/DTP date when it exists (Recidiviz/recidiviz-dashboards#8784)
In AZ, we want to show the "ACIS TPR date" criteria at the top of the criteria list. To do this, we override the requirements almost met (almost eligible criteria) in the opportunity class. Because of a too-broad early return, we were not doing this override in certain cases, so people who should have had the "ACIS TPR date" criteria overridden were instead showing up with their original criteria (including the "projected TPR date" item below).

GitOrigin-RevId: f4b1b6cf04f7ec3eaccea4a4930814268c436323