[Insights] Fix opportunity display in drilldown page (Recidiviz/recidiviz-dashboards#8796)
We inadvertently stopped displaying opportunities in the Insights opportunity drilldown page: we were displaying a new empty state when no search items were selected, thanks to Recidiviz/recidiviz-dashboards#8776, but technically there are no search items selected on the supervisor homepage opportunities drilldown page. (Is that what it's called? Is there a typical short name used for it?) I didn't want to add any supervisor homepage exceptions in the relatively complicated JSX for this component, so added a default on the presenter.

GitOrigin-RevId: 1fb37ca453b0ce56aaf1d0843622440d7a3d32ac