[Insights] Fix svg comments in RIR icons (Recidiviz/recidiviz-dashboards#8824)
Fix a bug where the new SVG files for the RIR modal were not rendering correctly because of a comment that looked like this: `<!-- ... --->`. The 3 hyphens at the end of the comment should be 2 hyphens, since -- seems to be not allowed after the start of a comment in SVG. The image now loads fine for me on Chrome.

GitOrigin-RevId: 5f0ed0055b5d575d499d3d4641bfd98c5c253ad9