[Tasks] Gate Tasks based on tasks route permission, not FV (Recidiviz/recidiviz-dashboards#8172)
Previously, we've assumed that users can access the /workflows/tasks page iff they have the workflowsSupervision route enabled. This PR makes the tasks and workflowsSupervision permissions completely separate. In other words, someone with the workflows but not tasks permission will not be able to access /workflows/tasks but will retain their access to all other workflows pages. After this PR and a follow-up PR, someone with the tasks permission but not workflows permission will only be able to access /clients, /residents, and /tasks routes.

Many of the changes in UserStore are formatting changes. The main changes here are to the functions userAllowedNavigation, isUserAllowedRoute, and getRoutePermission: adding a "subpage" parameter to isUserAllowedRoute and then changing userAllowedNavigation to account for that.

GitOrigin-RevId: 1d431e0b31ccfa42815c314a33a43a1c5a8c4d8a