[Insights] Base Vitals module permission on FV, not /operations access (Recidiviz/recidiviz-dashboards#8936)
We are going to be launching the Supervision Homepage Vitals module in Texas, where the Vitals product (/operations) is not launched. This PR changes the Vitals module permission to stop checking whether operations is an allowed route, instead only checking the supervisorHomepageVitals feature variant.

It looks like operations only exists as a route permission for users in Idaho and North Dakota, and all of the users in those states who have insights access can also access operations. So this PR will have no effect in existing states, while it'll allow TX users to see the Vitals module.

GitOrigin-RevId: eab51a100ea49fd7e5638e97c5761b0c6a7119ed