[Insights][Bug] Fallback to app metadata for user name in insights store (Recidiviz/recidiviz-dashboards#8967)
Sentry alert came via slack. In TN, we don't always receive the user name via the SSO link, so auth0 doesn't have that information. We don't want to throw an error, so we'll fall back on app metadata (which is sourced from the admin panel I believe). We'll remove the PII from the sentry error as well.

GitOrigin-RevId: f6f70308f517ed58e76a20bd480751f74aa82488