[Workflows] Disable save button while denial is being saved (Recidiviz/recidiviz-dashboards#8984)
Disable the denial view save button, until the denial submission has saved.

This bug was discovered during an Iowa trip - the team was working off of incredibly slow connection, which was necessary for the bug to surface. See Slack thread thread for more details. There is one previous instance of this happening in a different state in prod, but we never prioritized debugging.

Disabling the save button should prevent users from resubmitting while firestore is still syncing.

GitOrigin-RevId: db55f071aadd7e6c05313fec1d87bd5c3d21fdb9