[US_UT] Make Supervisor Homepage opportunities counts same as Workflows homepage (Recidiviz/recidiviz-dashboards#9164)
In UT, some testers were confused that the counts per officer in the SH Opportunities module were not the same as the counts seen when you click in. This is because the SH Opportunities module only counts "eligible" opportunities (not almost eligible, submitted, or denied), but the workflows homepage uses both eligible + almost eligible opportunities, or a custom function, for the count.

This PR changes the counts on the supervisor homepage so that
- if there is a custom counting function, all opportunities are counted based on the custom function
- otherwise, the counts are the same as before: only eligible opportunities are counted

The only opportunities with custom counting functions are UT Early Termination, MO Restrictive Housing, and OR Early Discharge, and only UT has a supervisor homepage launch imminent. So, this has the effect of correcting the counts for UT ET and keeping everything else the same. It also seems like a reasonable assumption for the future that if there is a custom way to count opportunities, we'd like to use it across product surfaces.

GitOrigin-RevId: 20ac485079487b41bb57a4f3ec70eb1028eb7330