[Tasks] Fix filter select (Recidiviz/recidiviz-dashboards#9256)
Recidiviz/recidiviz-dashboards#9149 changed the CaseloadTasksPresenterV2 constructor to select the first non-empty category on page load. This broke the Tasks table filters, which are managed by the same presenter, because the change made the constructor depend on the category counts. My understanding of the bug is: whenever you selected a filter, the counts in categories would change; due to the new dependency, this would cause mobx to create the presenter again, with the initial (empty) filter state, so the filter selection appeared to have no effect. This wasn't caught by tests because we only have Tasks tests of the presenter functions, not tests where we render components.

This PR changes the presenter to use a getter for selectedCategory.

GitOrigin-RevId: 39895415b75e228da5d61a81cf7ae82cb2f636d2