[Insights] Change 'N/A' placeholder to 'Unknown' in Supervisors List (Recidiviz/recidiviz-dashboards#9297)
This PR normalizes all unknown supervisor districts to "Unknown" for the supervisor list and supervisors pages. Now, all supervisors with an unknown district are grouped under "Unknown" at the end of the supervisor list. This makes a difference for states like Utah where some people were "N/A" and some people were "Unknown".

This means we can't distinguish between people who have a known-to-be-unknown vs. null district anymore, but according to Gina on Slack, it doesn't seem like we had a particularly strong reason to do that in the first place.

GitOrigin-RevId: 594bf3aeb620d8ff415af26f7d8122630718cdf4