[Workflows][US_IA] Tooltip for missing DNA submission (Recidiviz/recidiviz-dashboards#9362)
Adds a tooltip to the criteria related to missing DNA submissions (also added some typing to criteria functions).

This solves for the situation where officers may have submitted DNA to CODIS, but CODIS doesn't properly submit dates, so we can't see the submission in our data. We want officers to know that they should go ahead with submitting ED if they know that they submitted DNA to CODIS.

GitOrigin-RevId: 1d48b38cd27da99a2fcea1a96aa9fe7d8b7b6fe1