[Workflows][US_IA] Add a status to IA ED for Indefinite Snooze (Recidiviz/recidiviz-dashboards#9364)
dds SNOOZE_REVIEW client status to the IA ED opp, which will be used in the next PR for snoozes with supervisor approval.

In the clientStatus fn, the logic is now modified so that an officer DENIAL plus a supervisor DENIAL will either result in a ACTION_PLAN_REVIEW status (as usual) or a SNOOZE_REVIEW status (for indefinite snooze requests). We look at the denial reasons in the request to determine which status is relevant. A future PR will enforce that indefinite snooze denial reasons cannot be selected alongside other denials, so we wouldn't have collisions in the statuses.

* [Workflows][US_IA] Add a status to IA ED for Indefinite Snooze

* Add menu button config and test

* use public safety const

* add tab for status

* rework the clientStatus tab for readability

* Throw error for expected unreachable state

* update test

GitOrigin-RevId: 0c66ec3d5aed9a05c65cfb0528ce241813271eee