[Workflows][US_IA] Send indefinite snooze requests to the supervisor approval tab (Recidiviz/recidiviz-dashboards#9358)
Updates the OpportunityDenialView to send snooze requests to the "Supervisor Review" tab, rather than immediately snoozing the client.

This PR:

Adds indefiniteDenialReasons field to the opportunity class - subclasses can override with their opportunity's indefinite reasons. We override this in IA to return the COURT and INTERSTATE(IC-IN) reason. This list of reasons will be used in the OpportunityDenialView (see next bullet).
Adds related class tests
Updates the OpportunityDenialView logic - If the selected denial reasons include an indefinite reason, we call setOfficerAction and create an officer DENIAL action rather than immediately snoozing the client.
Adds a relevant toast message and updates the IA sidepanel component with an indefinite snooze string
I have been testing this locally and I think all statuses and actions from various tabs are aligned with expected behavior! This means the remaining work for indefinite snooze after this lands is Recidiviz/recidiviz-dashboards#9163 (the UI work).

GitOrigin-RevId: 2802d762f8290121b49950e897c0bd0e7b42e7db