[Workflows] Add section component for snooze UI (Recidiviz/recidiviz-dashboards#9398)
Creates a DenialReasonSection component so we can split indefinite reasons from temporary reasons in a future PR.

This is a refactor to make future PRs simpler:

Moves reason list + heading into its own component
Pulls out the handler function for selecting a reason into a const (will be passed to multiple DenialReasonSection components in future)
Moves the selection handler to be called in the Checkbox onChange() instead of the MenuItem (this clears the way for disabling denial reasons). The MenuItem onChange was at odds with the checkbox behavior so we had to override the checkbox via e.preventDefault previously. However that is preventing us from using the disabled checkbox behavior. See this old bug fix thread for more details.

GitOrigin-RevId: 2b292bfb511ac325e2ad9a71010cf7fd6d0f815c