[Workflows] Disable and de-select other snooze reasons when selecting indefinite reason (Recidiviz/recidiviz-dashboards#9416)
Disable/deselect other reasons when selecting an indefinite snooze reason.

Indefinite snooze reasons are meant to be selected exclusive of other reasons. To enforce this, we'll disable and de-select other reasons when selecting an indefinite reason. This PR adds state to track the currently disabled reasons, so that we can toggle the relevant checkboxes to and from disabled as an indefinite reason is selected or de-selected. Also adds styling logic to gray out the text of disabled reasons.

To test via the preview, go to Iowa ED, and select an indefinite reason (COURT, or INTERSTATE (IC-IN)).

GitOrigin-RevId: 5cc860d12d625e114310a5209e64ba3fc6b70e5a