[Workflows][US_IA] Add indefinite snooze section to denial view (Recidiviz/recidiviz-dashboards#9399)
Visually separates indefinite snooze reasons from temporary snooze reasons.

This PR:

Adds a new DenialReasonSection for indefinite reasons if they exist.
Filters out indefinite reasons from the original DenialReasonSection
Adds the title field for the original DenialReasonSection to the opportunity so we can override it in Iowa when snooze is enabled

GitOrigin-RevId: f6e281bd273e1de41479662289a29306d794e224