[Workflows] Add UsUtDates tests and fix empty Milestones in sidebar (Recidiviz/recidiviz-dashboards#9344)
This change creates a test file for UsUtDates component which tests the following:

The report due date should be hidden when it's in the past and the person was eligible for ET.
When the halfTimeDate is present, it should take precedence over the date calculated based on the start and end dates.

It also fixes a bug where the Milestones component was returning an empty div instead of null, causing two dividers to be displayed in the sidebar.

GitOrigin-RevId: 4756d9ce16dacd0e51cba55a474293aeaff7c0cf