[Workflows][Snooze] Make Opp Config the source of truth for denial reason snooze lengths  (Recidiviz/recidiviz-dashboards#9535)
Adds and modifies relevant logic on the Opportunity Configuration and Opp Base classes so that varied snooze lengths (and indefinite snoozes) can be set via the Opp Config class.

In the future, we'd like this to be admin panel configurable (see Recidiviz/recidiviz-dashboards#9350), but for now, this change should make it simple/quick for a SWE to go in and set custom snooze lengths per denial reason via a custom Opp Config class.

This PR:

Adds a maxSnoozeDaysByDenialReason mapping on the OpportunityConfiguration
a. by default, this will map every denial reason to the opportunity's snooze config's maxSnoozeDays value.
b. an opportunity's custom Opp Config can then override the mapping to set unique values per denial reason.
Adds an indefiniteDenialReasons field on the OpportunityConfiguration
a. This looks at the maxSnoozeDaysByDenialReason mapping, and returns the denial reasons that have been mapped to an undefined max snooze.
b. The opportunity class's indefiniteDenialReasons field now can just pull indefinite denial reasons from the configuration.
The maxManualSnoozeDays on the Opportunity class (this function actually determines the snooze slider's max value based on selected denial reasons) has been updated to look at the configured maxSnoozeDaysByDenialReason mapping.
Updates the custom IA configuration class to override maxSnoozeDaysByDenialReason for the relevant denial reasons
Adds tests for the OpportunityConfiguration class implementation and the IA configuration for maxSnoozeDaysByDenialReason and indefiniteDenialReason

GitOrigin-RevId: ed415e5c0cdb6a3f905db93bb397230b2b2774fb