[Workflows][US_IA][Snooze Features] Add userInput field to firestore types and functions (Recidiviz/recidiviz-dashboards#9602)
Adds the userInput field to the denial and officerAction types so that we can start surfacing/storing fee information for Iowa

In Iowa, we'd like users to start including the outstanding fee amount when snoozing someone for the "Fines & Fees" reason. We want to store that fee amount (perhaps for future analysis) and display it in the sidepanel.

In order to do this, we'll be adding a userInput mapping onto the denial blob in firestore. This field will map denial reasons to the user input at the time of snoozing. See design doc "firestore schema" section here.

This is PR 1/2 to enable text input for denial reasons. Future PRs will handle configuring denial reasons to require text input and the UI changes.

This PR covers the firestore related functions and types:

adds userInput params to updateOpportunityDenial() fn so we can set and unset the field on the denial blob
updates firestore OfficerDenialAction and Denial type to include the userInput field
updates setDenialReasons on the opp class to take in the userInput param.
Adds tests for setDenialReasons and updateOpportunityDenial

GitOrigin-RevId: acd96395f038fd9decc73f1769535abf7ed5eb77