[Workflows][US_IA][Snooze features] Add configurable text input fields to denial view (Recidiviz/recidiviz-dashboards#9614)
Updates the denial view to add text input fields for denial reasons when configured.

In Iowa, we'd like users to start inputting the outstanding fee amount when snoozing someone for the "Fines & Fees" reason. We want to store fee amount (perhaps for future analysis) and display it in the sidepanel.

This PR enables us to configure which denial reasons are associated with a text input field and to specify characteristics of the text field like input type and copy. Because there are existing denial reasons that require text input (e.g. "public safety risk" in IA, and the "Other" reason) and likely will be more in the future, we've made this a field on the Opp Config so different states can also customize their opps.

Configured denial reason will trigger a text box for user input when selected. The user input will be saved to firestore when the selected denial reasons are saved.

This is PR 2/2 to enable text input for denial reasons. I broke everything up by commit so reviewing by commit is a good option. Big changes are:

GitOrigin-RevId: 4a8a9418a19abbfa168ffa8cbf0de3537d1fbb39