[Tasks][Route Planner] Make new page at /workflows/tasks/home-contact-route-planner (Recidiviz/recidiviz-dashboards#9627)
This PR creates a new page at /workflows/tasks/home-contact-route-planner for the Route Planner project.

When adding the new path to WORKFLOWS_PATHS, I noticed we seemed to be conflating our internal identifiers for pages with the strings that we would want to appear in a URL section, so I tried to make the types more clear. This is the first commit.

Access to the page is controlled by the tasksRoutePlanner feature variant, which is currently only enabled in TX in staging for Recidiviz users. A FV is used instead of a route permission because we will eventually want to alter other pages' layout based on whether someone has access to this page.

This PR also changes parseLocation to account for a workflows URL with multiple parts of the path where the third part of the path isn't a person ID (like in /workflows/clients/someid123).

GitOrigin-RevId: 11b964b13ed246bc06c6c22f6b2ac974fe6c83b2