[Tasks][Route Planner] Set up client list skeleton (Recidiviz/recidiviz-dashboards#9638)
This PR sets up the basic desktop version of the route planner page at /workflows/tasks/home-contact-route-planner (only available in Texas).

To do so, introduce a new RoutePlannerPresenter class that will later handle state for the overall page and the map, which also manages a RoutePlannerClientsPresenter. This latter presenter handles hydration for the highlighted client select portion of the page so that, when new caseloads are selected, only the highlighted portion will reload, not the entire page.

GitOrigin-RevId: 4ff853c3b68ad27a7b29ddc118c04a7091fe0722