[Tasks][Route Planner] Enable selecting/deselecting clients (Recidiviz/recidiviz-dashboards#9663)
Adds click interactions to the cards introduced in Recidiviz/recidiviz-dashboards#9644. The RoutePlannerClientsPresenter will keep track of the list of selected clients, and the ClientCard component will display the client's position in the list. Also, when the user removes an officer from their search bar, any clients assigned to those officers will be removed from the user selections.

GitOrigin-RevId: ce00adbf10a5690a92538ac456e36ea5a6e5a867