[Workflows][US_IA] Multiselect reason for Iowa indefinite snooze (Recidiviz/recidiviz-dashboards#9819)
In Iowa, we want to allow multi-select of denial reasons even when one of those reasons is an indefinite reason.

We're not making this configurable as Claycia is the state-agnostic PM for this and her perspective is that indefinite reasons should not be combined with non-indefinite reasons.

Slack thread(s) here (responses weren't threaded). We'd like to get this in for the Oct 2 launch/training push.

Tested that in Iowa ED you can now multiselect and the snooze approval flow works as expected. Confirmed that other state with indefinite snooze is unaffected (MO work release).

GitOrigin-RevId: 20c6805e29a25762af4b99bd5cb661277cb71c37