[Workflows][Supervisor Approvals] Add new "Supervisor Review" tab, statuses, and menu options (Recidiviz/recidiviz-dashboards#9772)
Updates the opportunity configuration, opportunity base class, and the menu UI to add state-agnostic supervisor review functionality.

This PR is organized by commit. A summary of all the changes is as follows:

Adds fields to the opp config related to the supervisor approval tab and status copy
Updates the OpportunityBase class to pull in fields we need for supervisor approvals
a couple firestore functions that had only lived on the IA ED opp needed to be on the base class (deleteActionHistory, setSupervisorResponse
Modification of the IA ED opportunity class and config to override new configurable fields (so there's no change in UI for IA users)
Updates the state agnostic MenuButton component to use approval/denial menu items when the opportunity is under review
This button is getting a bit crowded so I've added a TODO to look into a presenter class.
Updates the OppDenialView component to handle all supervisor approvals, not just indefinite snooze approvals.
Updates the "Revert changes" button logic to delete action history and to display in the supervisor review tab
Adds tests for the tabTitle field
Adds tests for the supervisor response firestore function
Confirmed that IA behavior has not change in the preview URL and all existing IA opp and config tests pass. The next PR updates MI opps to enable supervisor review for indefinite snooze so that preview URL is a good place to manually test the new state-agnostic implementation.

GitOrigin-RevId: dfca32d52d42deba87e0278431dbf39033a08ee8