[Workflows][US_ME] Calculate SCCP time until release consistently (Recidiviz/recidiviz-dashboards#10270)
A user in Maine reported that the number of months until release that we were showing in a SCCP criteria item did not agree with the release date we display.

The cause: we were calculating the release date based on the eligible date in the criteria item, and erroneously adding 6 months to the difference. (IIRC I added the 6 months to compensate for the eligible date equaling 24 months before release date at the time, but it can be either 24 or 30 depending on statewide caseload sizes.) This roundabout calculation of release date was necessary because, as ME SCCP is an older opportunity, criteria hydration only had access to the fields from the criteria item, not the full opportunity.

The fix: This PR changes ME SCCP to use admin panel-defined criteria copy in addition to the hardcoded criteria in UsMeSCCPOpportunity.ts. Moving the usMeXMonthsRemainingOnSentence item to the admin panel config and using the newer criteria hydration lets us calculate the months until release directly from the resident's release date. This PR also changes the "almost eligible status message" to use the same helper fn so our calculation is consistent with the release date in both places.

Ideally we would get rid of all of the criteria copy still in code, but I just moved over the one criteria item in the interest of having a more specific PR and not spending too much time on this. I'll promote the new admin panel SCCP config to prod if this approach seems reasonable.

GitOrigin-RevId: 2a8906e214d70d9508b1b07f66af0af1cc8aaa1d