[Tasks][Route Planner] Add buttons to copy map link and email to self (Recidiviz/recidiviz-dashboards#10306)
In the TX Home Contact Route Planner, add buttons to the info box overlay over the map to copy the Google Maps link to clipboard and email the maps link to oneself.

The maps link is formatted as a Google Maps "Directions URL".
Emailing oneself uses the API endpoint introduced in [Workflows] Create external request endpoint for emailing the logged-in user recidiviz-dataRecidiviz/recidiviz-dashboards#48593
The map uses the copy in the Route Planner PRD.

GitOrigin-RevId: 7caadb9ef1b147f90f1594620901210309c5222d