[Tasks][Route Planner] Use Google Maps Geocoding API for addresses (Recidiviz/recidiviz-dashboards#10449)
For the Home Contact Route Planner, we want to prevent users from selecting people who have problematic ("bad") addresses for the Google Maps embed (i.e. ambiguous addresses or addresses containing typos). This PR introduces a call to Google Maps's Geocoding API when a user selects a client. The Geocoding API takes in addresses and returns a "place ID", an identifier for a specific place in Google Maps that can be used in the embed URL. If the API gives us back no results or multiple results, we stop the user from selecting the client. The bulk of the changes are in the RoutePlannerClientsPresenter.

In order to limit the number of requests we send to this API, we also store the place IDs in Firestore in clientUpdatesV2. We read from the Firestore collection when displaying cards, so if we've already tried to geocode someone's address and it didn't work, we can make them unselectable from the start.

GitOrigin-RevId: 11e0fedcee98be5a50328453a29a7faa05b92483