[Tasks][Route Planner] Display scheduled contact info in Route Planner (Recidiviz/recidiviz-dashboards#10475)
Pulls in contact dates for upcoming appointments to the Route Planner. Tasks with upcoming appointments have pills that are formatted differently. (Somewhat confusingly, we're internally calling these dates "scheduled contact dates", but the dates can apply to either "scheduled home" or "unscheduled home" contacts. The latter names refer to whether or not the client knows the contact is happening in advance, but the officer can make an appointment in OIMS for any contact.)

As part of this work, I made a new file RoutePlannerClientCard for the component that handles displaying one client, since it was getting a bit long.

GitOrigin-RevId: 94cdd5a864a39ad94b49720e255ba11d6f4fc23d