[Tasks][Route Planner] Fix map display, outgoing map links (Recidiviz/recidiviz-dashboards#10552)
This PR adjusts the map display in the Home Contact Route Planner to fix bugs / address bug bash requests:

When the user has not selected any clients, shows a map of Texas and a prompt to select clients, instead of trying to plot the starting location (which didn't work when the user hadn't selected a starting location).
The placeholder starting location is now the address of the officer on the first client of the route, instead of the address of the first selected officer. (In practice we don't expect these to differ often if at all, but it seems more correct.)
The map now displays a one-way route starting from the user-selected starting location, instead of going back to the starting location at the end.
Fixes some broken external links I introduced in `[Tasks][Route Planner] Use Google Maps Geocoding API for addresses`. Some of the google maps APIs we use accept place IDs in place of addresses, but some of them accept them in a different query parameter.
Adds a button to open the route link in a new tab. This is not in the PRD but I found helpful for debugging ^ and it seems helpful in general.

GitOrigin-RevId: db867c9dc687215f1fc131299a1488050bbe85a6