[Deploy] Automatically reauth to gcloud ADC/Firebase CLI (Recidiviz/recidiviz-dashboards#10700)
This PR adds a check for gcloud ADC and Firebase CLI credentials to the start of the deploy script, and automatically opens a browser window to try to reauth the user if the credentials have expired. It's a little annoying to have to remember to run the reauth commands each time, and the deploy script sometimes gives unclear error messages when it proceeds without these credentials.

GitOrigin-RevId: 0e7a374a4618db79f430a2ddc971dbd2f914e70f