[US_TX][Workflows] stop marking ARS/ERS as pending on form download (Recidiviz/recidiviz-dashboards#10749)
We're going to change the submitted tab title for Texas's ARS/ERS opportunities to something like "submitted for review", because officers found "pending" to be confusing. Accordingly, this PR stops us from moving people to the submitted tab on form download - the reasoning being that a form might be downloaded but not yet submitted to the officers' higher-ups.

GitOrigin-RevId: 5fe67993d2b7d9e3080a440670938293cff0f25d