[US_NC][JII] Add buttons to navigate between form pages (Recidiviz/recidiviz-dashboards#11002)
This PR sets up the assessment entry screen for the North Carolina RNA tablet app at UsNcRNALanding with placeholder routing so that the resident home page redirects to this landing page. It also creates a new component NavigationButtons for the buttons forward and back among pages of questions within the assessment.

GitOrigin-RevId: 315bbff32706dd96980318b5d7003fe50589533e