[JII][US_NC] Add RNA copy object and use it on form page (Recidiviz/recidiviz-dashboards#11120)
This PR adds some copy and configuration for the RNA form in a new file called usNcRNAFormSpec.

All of the actual copy is in nested objects in case we ever want to move over to the translation framework, but I thought that would be too heavyweight for right now since we are not planning to support multiple languages. Based on Ian's guidance, the copy and configuration are in separate objects.

GitOrigin-RevId: ffd336df6e765be1f912979504da765110135294